refactor: moved e-signet related methods to esignet-api.ts
